driver: make grid and block dims 32 bits

+++ gluon-sim/src/common/base.rs
@@ -109,8 +109,8 @@ pub struct MemResp {
 #[derive(Debug, Default, Clone)]
 pub struct ThreadBlocks {
     pub pc: u32,
-    pub block_idxs: Vec<(u16, u16, u16)>,
-    pub block_dim: (u16, u16, u16),
+    pub block_idxs: Vec<(u32, u32, u32)>,
+    pub block_dim: (u32, u32, u32),
     pub regs: u32,
     pub shmem: u32,
 }
